+++ dbt/adapters/obmysql/__init__.py
@@ -12,12 +12,12 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 from dbt.adapters.base import AdapterPlugin
-from dbt.adapters.oceanbase_mysql.connections import OBMySQLCredentials
-from dbt.adapters.oceanbase_mysql.impl import OBMySQLAdapter
-from dbt.include import oceanbase_mysql
+from dbt.adapters.obmysql.connections import OBMySQLCredentials
+from dbt.adapters.obmysql.impl import OBMySQLAdapter
+from dbt.include import obmysql
 
 Plugin = AdapterPlugin(
     adapter=OBMySQLAdapter,
     credentials=OBMySQLCredentials,
-    include_path=oceanbase_mysql.PACKAGE_PATH,
+    include_path=obmysql.PACKAGE_PATH,
 )
