+++ dbt/adapters/obmysql/column.py
