+++ dbt/include/obmysql/dbt_project.yml
@@ -1,5 +1,5 @@
 config-version: 2
-name: dbt_oceanbase_mysql
+name: dbt_obmysql
 version: 1.0
 
 macro-paths: ["macros"]
