+++ dbt/include/obmysql/macros/adapters.sql
@@ -1,23 +1,23 @@
-{% macro oceanbase_mysql__create_schema(relation) -%}
+{% macro obmysql__create_schema(relation) -%}
   {%- call statement('create_schema') -%}
     create database if not exists {{ relation.without_identifier().include(schema=False) }}
   {%- endcall -%}
 {% endmacro %}
 
-{% macro oceanbase_mysql__drop_schema(relation) -%}
+{% macro obmysql__drop_schema(relation) -%}
   {%- call statement('drop_schema') -%}
     drop database if exists {{ relation.without_identifier().include(schema=False) }}
   {%- endcall -%}
 {% endmacro %}
 
-{% macro oceanbase_mysql__list_schemas(database) %}
+{% macro obmysql__list_schemas(database) %}
   {% call statement('list_schemas', fetch_result=True, auto_begin=False) %}
     show databases
   {% endcall %}
   {{ return(load_result('list_schemas').table) }}
 {% endmacro %}
 
-{% macro oceanbase_mysql__list_relations_without_caching(schema_relation) %}
+{% macro obmysql__list_relations_without_caching(schema_relation) %}
   {% call statement('list_relations_without_caching', fetch_result=True) -%}
     select
       '{{ schema_relation.database }}' as `database`,
@@ -46,7 +46,7 @@
   {{ return(load_result('list_relations_without_caching').table) }}
 {% endmacro %}
 
-{% macro oceanbase_mysql__get_empty_schema_sql(columns) %}
+{% macro obmysql__get_empty_schema_sql(columns) %}
     {%- set col_err = [] -%}
     select
     {% for i in columns %}
@@ -62,7 +62,7 @@
     {%- endif -%}
 {% endmacro %}
 
-{% macro oceanbase_mysql__get_create_index_sql(relation, index_dict) -%}
+{% macro obmysql__get_create_index_sql(relation, index_dict) -%}
   {%- set index_config = adapter.parse_index(index_dict) -%}
   {%- set comma_separated_columns = ", ".join(index_config.columns) -%}
   {%- set index_name = index_config.get_name(relation) -%}
@@ -85,18 +85,18 @@
   {% endif %};
 {%- endmacro %}
 
-{% macro oceanbase_mysql__copy_grants() %}
+{% macro obmysql__copy_grants() %}
     {{ return(False) }}
 {% endmacro %}
 
-{% macro oceanbase_mysql__alter_relation_comment(relation, comment) %}
+{% macro obmysql__alter_relation_comment(relation, comment) %}
   {%- if not relation.is_table -%}
     {{ exceptions.raise_fail_fast_error(relation.type ~ " do not support setting comment") }}
   {%- endif -%}
   alter table {{ relation }} set comment='{{ comment }}';
 {% endmacro %}
 
-{% macro oceanbase_mysql__rename_relation(from_relation, to_relation) -%}
+{% macro obmysql__rename_relation(from_relation, to_relation) -%}
   {#
     2-step process is needed:
     1. Drop the existing relation
@@ -110,22 +110,22 @@
   {% endcall %}
 {% endmacro %}
 
-{% macro oceanbase_mysql__get_show_indexes_sql(relation) %}
+{% macro obmysql__get_show_indexes_sql(relation) %}
     show index from `{{ relation.identifier }}` from `{{ relation.database }}`
 {% endmacro %}
 
-{% macro oceanbase_mysql__list_indexes(relation) %}
+{% macro obmysql__list_indexes(relation) %}
     {% call statement('list_indexes', fetch_result=True, auto_begin=False) %}
         {{ get_show_indexes_sql(relation) }}
     {% endcall %}
     {{ return(load_result('list_indexes').table) }}
 {% endmacro %}
 
-{%- macro oceanbase_mysql__get_drop_index_sql(relation, index_name) -%}
+{%- macro obmysql__get_drop_index_sql(relation, index_name) -%}
     drop index `{{ index_name }}` on {{ relation }}
 {%- endmacro -%}
 
-{% macro oceanbase_mysql__get_columns_in_relation(relation) -%}
+{% macro obmysql__get_columns_in_relation(relation) -%}
   {% call statement('get_columns_in_relation', fetch_result=True) %}
       select
         column_name,
