+++ dbt/include/obmysql/macros/relations/incremental/merge.sql
@@ -1,4 +1,4 @@
-{% macro oceanbase_mysql__get_delete_insert_merge_sql(target, source, unique_key, dest_columns, incremental_predicates) -%}
+{% macro obmysql__get_delete_insert_merge_sql(target, source, unique_key, dest_columns, incremental_predicates) -%}
 
     {%- set dest_cols_csv = get_quoted_csv(dest_columns | map(attribute="name")) -%}
