+++ dbt/include/obmysql/macros/relations/materialized_view/drop.sql
@@ -1,3 +1,3 @@
-{% macro oceanbase_mysql__drop_materialized_view(relation) -%}
+{% macro obmysql__drop_materialized_view(relation) -%}
     drop materialized view if exists {{ relation.include(schema=False) }}
 {%- endmacro %}
