+++ dbt/include/obmysql/macros/relations/table/drop.sql
@@ -1,3 +1,3 @@
-{% macro oceanbase_mysql__drop_table(relation) -%}
+{% macro obmysql__drop_table(relation) -%}
     drop table if exists {{ relation.include(schema=False) }}
 {%- endmacro %}
