+++ dbt/include/obmysql/macros/relations/view/drop.sql
@@ -1,3 +1,3 @@
-{% macro oceanbase_mysql__drop_view(relation) -%}
+{% macro obmysql__drop_view(relation) -%}
     drop view if exists {{ relation.include(schema=False) }}
 {%- endmacro %}
