+++ dbt/include/obmysql/macros/snapshot/strategies.sql
@@ -1,4 +1,4 @@
-{% macro oceanbase_mysql__snapshot_hash_arguments(args) -%}
+{% macro obmysql__snapshot_hash_arguments(args) -%}
   md5({%- for arg in args -%}
     coalesce(cast({{ arg }} as char ), '')
       {% if not loop.last %} || '|' || {% endif %}
