+++ dbt/include/obmysql/macros/timestamp.sql
@@ -1,0 +1,3 @@
+{% macro obmysql__current_timestamp() -%}
+  now()
+{%- endmacro %}
