+++ dbt/include/obmysql/profile_template.yml
@@ -1,5 +1,5 @@
 fixed:
-  type: oceanbase_mysql
+  type: obmysql
 prompts:
   host:
     hint: 'hostname for the instance'
