--- dbt/include/oceanbase_mysql/macros/timestamp.sql
@@ -1,3 +1,0 @@
-{% macro oceanbase_mysql__current_timestamp() -%}
-  now()
-{%- endmacro %}
