+++ tests/functional/materializations/test_incremental.py
@@ -18,8 +18,8 @@ import pytest
 from dbt.adapters.base import BaseAdapter, BaseRelation
 from dbt.adapters.base.column import Column as BaseColumn
 from dbt.adapters.contracts.relation import RelationType
-from dbt.adapters.oceanbase_mysql.column import OBMySQLColumn
-from dbt.adapters.oceanbase_mysql.relation import OBMySQLRelation
+from dbt.adapters.obmysql.column import OBMySQLColumn
+from dbt.adapters.obmysql.relation import OBMySQLRelation
 from dbt.cli.main import dbtRunner, dbtRunnerResult
 from dbt.tests.adapter.column_types.test_column_types import run_dbt
 from tests.functional.utils import BaseOBMySQLTestCase
