+++ tests/functional/materializations/test_materialized_view.py
@@ -18,9 +18,9 @@ from dbt_common.exceptions import DbtRuntimeError
 
 from dbt.adapters.base import BaseAdapter, BaseRelation
 from dbt.adapters.contracts.relation import RelationType
-from dbt.adapters.oceanbase_mysql import OBMySQLAdapter
-from dbt.adapters.oceanbase_mysql.impl import OBMySQLIndex
-from dbt.adapters.oceanbase_mysql.relation import OBMySQLRelation
+from dbt.adapters.obmysql import OBMySQLAdapter
+from dbt.adapters.obmysql.impl import OBMySQLIndex
+from dbt.adapters.obmysql.relation import OBMySQLRelation
 from dbt.tests.adapter.column_types.test_column_types import run_dbt
 from tests.functional.utils import BaseOBMySQLTestCase
