+++ tests/functional/utils.py
@@ -13,7 +13,7 @@
 # limitations under the License.
 import pytest
 
-from dbt.adapters.oceanbase_mysql.connections import (
+from dbt.adapters.obmysql.connections import (
     OBMySQL_DIALECT_TYPE,
     OBMySQLCredentials,
 )
